+++ core/timestepping/TimeStepping.cpp
@@ -42,23 +42,13 @@ void ITimeStepping::step(IScheduler& scheduler, ISolver& solver, Statistics& sta
     this->stepParticles(scheduler, solver, stats);
 
     // kick & drift attractors
-    Array<Size> toRemove;
     for (Attractor& a : storage->getAttractors()) {
         a.velocity += a.acceleration * timeStep;
         a.position += 0.5_f * a.velocity * timeStep;
 
-        // process black holes
-        if (a.settings.getOr<bool>(AttractorSettingsId::BLACK_HOLE, false)) {
-            /// \todo parallelize
-            ArrayView<const Vector> r = storage->getValue<Vector>(QuantityId::POSITION);
-            for (Size i = 0; i < r.size(); ++i) {
-                if (getSqrLength(a.position - r[i]) < sqr(a.radius)) {
-                    toRemove.push(i);
-                }
-            }
-        }
+        // process particle interactions
+        a.interact(scheduler, *storage);
     }
-    storage->remove(toRemove, Storage::IndicesFlag::INDICES_SORTED | Storage::IndicesFlag::PROPAGATE);
 
     // update time step
     CriterionId criterionId = CriterionId::INITIAL_VALUE;